% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_ify.R
\name{c_ify}
\alias{c_ify}
\title{Print a vector formatted as valid R code for a vector}
\usage{
c_ify(
  s,
  from = c("r", "r_char", "stata", "python", "py_char"),
  from_start = NULL,
  from_sep = NULL,
  from_end = NULL,
  string = from \%in\% c("r_char", "py_char", "stata"),
  to_start = NULL,
  to_sep = NULL,
  to_end = NULL,
  trim = TRUE,
  silent = FALSE
)
}
\arguments{
\item{s}{A string}

\item{from}{The formatting of the origin string. This can be \code{'stata'} or \code{'r'} (\code{a b c}), \code{'r_char'} (\code{"a" "b" "c"}), \code{'python'} (\code{[a, b, c]}), or \code{'py_char'} (\code{['a', 'b', 'c']}). Any format can be supported using custom \code{from_} options below.}

\item{from_start}{Text at the beginning of the string to remove before separating the elements. Or just give a number to remove that many characters.}

\item{from_sep}{Character(s) separating each element.}

\item{from_end}{Text at the end of the string to remove before separating the elements.  Or just give a number to remove that many characters.}

\item{string}{Set to \code{TRUE} to set the \code{to_} defaults to produce an R character vector, or \code{FALSE} to set the \code{to_} defaults to produce an R numeric vector.}

\item{to_start}{Text at the beginning of the output string.}

\item{to_sep}{Character(s) separating each element in the output.}

\item{to_end}{Text at the end of the output string.}

\item{trim}{Merge all multiple-whitespaces into single whitespaces before separating the elements, handy if \code{from_sep = ' '}.}

\item{silent}{Suppress the message to console and just return the result as a string.}
}
\description{
This function takes a string, such as \code{"a b c"}, and returns code that will produce that vector in R, such as \code{c("a","b","c")}.
}
\details{
This is handy when copy/pasting vectors from R console output or other languages (like Stata) into your code.
}
\examples{

c_ify('1 2 3')
c_ify('[1, 2, 3]', from = 'python')
c_ify('var1 var2 var3', from = 'stata')
c_ify('\%\%\%a_b_c$$$', from_start = '\%\%\%', from_sep = '_', from_end = '$$$', string = TRUE)

}
